backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
openai
qdrant-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
sqlalchemy
psycopg2-binary
python-dateutil
pydantic
python-dotenv
pandas
numpy
pydantic[email]
loguru
